Updated examples with new keywords
