+++ example/keyring_example/subconfigs/model.yaml
@@ -1,3 +1,3 @@
-_target_: main.MyModel
+_instance_: main.MyModel
 layer_size: 128
 activation: relu
