+++ example/nested_unlocks_example/config1.yaml
@@ -1,6 +1,6 @@
 ?debug: true
 epochs: 128
 model:
-  _target_: main.MyModel
+  _instance_: main.MyModel
   layer_size: 128
   activation: relu
