+++ example/partial_instantiate_example/config.yaml
@@ -1,5 +1,5 @@
 neural_network:
-  _target_: "main.SuperCoolNeuralNetwork"
+  _partial_: "main.SuperCoolNeuralNetwork"
   hidden_size: 128
   activation: "tanh"
   dropout: 0.2
