+++ example/profiles_example/models/debug/head.yaml
@@ -1,2 +1,2 @@
 head_model:
-  _target_: models.debug_head
+  _instance_: models.debug_head
