+++ example/profiles_example/models/train/backbone.yaml
@@ -1,2 +1,2 @@
 backbone_model:
-  _target_: models.backbone
+  _instance_: models.backbone
